# docs/openapi.yaml
openapi: 3.0.0
info:
  title: WOTLWEDU API
  version: 0.0.49
paths:
...
  /v1/organization/:
    get:
      summary: GET /v1/organization/
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/organization/
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/tutorial/poll:
    get:
      summary: GET /v1/tutorial/poll
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/picture/:
    get:
      summary: GET /v1/picture/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/picture/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/preference/:
    get:
      summary: GET /v1/preference/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/preference/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/notification/:
    post:
      summary: POST /v1/notification/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    get:
      summary: GET /v1/notification/
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: items
        in: query
        required: false
        schema:
          type: integer
      - name: statusId
        in: query
        required: false
        schema:
          type: integer
      - name: type
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/register/:
    post:
      summary: Register a consumer account with a personal organization and first space
      parameters: []
      responses:
        '200':
          description: Registration accepted and confirmation email queued
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - firstName
              - password
              properties:
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                password:
                  type: string
                  minLength: 8
                spaceName:
                  type: string
                  description: Optional display name for the first personal space.
                organizationName:
                  type: string
                  description: Optional display name for the personal organization.
      description: |
        Self-service consumer registration creates a private organization, creates
        the first personal space, makes the new user the admin for that space, and
        starts the poll tutorial. The account remains inactive until the email
        confirmation token is used.
...
  /v1/item/:
    get:
      summary: GET /v1/item/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/item/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/vote/:
    get:
      summary: GET /v1/vote/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/vote/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/list/:
    get:
      summary: GET /v1/list/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/list/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/login/:
    post:
      summary: POST /v1/login/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/person/:
    get:
      summary: GET /v1/person/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/person/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
...
  /v1/circle/:
    post:
      summary: POST /v1/circle/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    get:
      summary: GET /v1/circle/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/capability/:
    get:
      summary: GET /v1/capability/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/capability/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
  /v1/space/:
    get:
      summary: GET /v1/space/
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
    post:
      summary: POST /v1/space/
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/poll/:
    post:
      summary: POST /v1/poll/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    get:
      summary: GET /v1/poll/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/role/:
    post:
      summary: POST /v1/role/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    get:
      summary: GET /v1/role/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
  /v1/category/:
    post:
      summary: POST /v1/category/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    get:
      summary: GET /v1/category/
      parameters: []
      responses:
        '200':
          description: Successful response
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        alias:
          type: string
        imageId:
          type: string
        email:
          type: string
        creator:
          type: string
        lastLogin:
          type: string
        active:
          type: boolean
        admin:
          type: boolean
        protected:
          type: boolean
        verified:
          type: boolean
        enable2fa:
          type: boolean
        secret2fa:
          type: string
        token2fa:
          type: string
        auth:
          type: string
        resetToken:
          type: string
        resetTokenExpire:
          type: string
        refreshToken:
          type: string
        refreshTokenExpire:
          type: string
        registerToken:
          type: string
        registerTokenExpire:
          type: string
        changeToEmail:
          type: string
        updatedAt:
          type: string
...